Events and states movements; test support for states movement on event; junit 5 added; event/state classes package reorganization
