+++ src/main/kotlin/edu/shamalov/os/OperatingSystem.kt
@@ -1,6 +1,5 @@
 package edu.shamalov.os
 
-import edu.shamalov.os.event.Event
 import edu.shamalov.os.schedule.Scheduler
 
 interface OperatingSystem {
