+++ src/main/kotlin/edu/shamalov/os/Task.kt
@@ -1,8 +1,5 @@
 package edu.shamalov.os
 
-import edu.shamalov.os.event.Event
-import edu.shamalov.os.state.State
-
 interface Task {
     val isBasic: Boolean
     val isCompleted: Boolean // TODO: do we need this? Looks like spec do not follow it
